requirements.txt
pybullet==3.1.7
scipy
numpy==1.22.3
gymnasium==0.28.1
yamldataclassconfig==1.5.0
nptyping<2.0